Conflicts:
	surface_src/java_src/Alice/src/com/roboclub/robobuggy/main/config.java
	surface_src/java_src/Alice/src/com/roboclub/robobuggy/main/mainFile.java
	surface_src/java_src/Alice/src/com/roboclub/robobuggy/messages/DateFormatUtils.java

+++ surface_src/java_src/Alice/src/com/roboclub/robobuggy/messages/WheelAngleCommand.java
@@ -4,15 +4,25 @@ import java.util.Date;
 
 import com.roboclub.robobuggy.ros.Message;
 
+/**
+ * @author ?
+ *
+ * @version 0.5
+ * 
+ * CHANGELOG: NONE
+ * 
+ * DESCRIPTION: TODO
+ */
+
 // Represents raw measurement from the IMU
 public class WheelAngleCommand implements Message {
 	public static final String version_id = "autonomous_angleV0.0";
 
 	public Date timestamp;
-	public double angle;
+	public byte angle;
 
 	// Makes an encoder measurement with the time of Now.
-	public WheelAngleCommand(double angle) {
+	public WheelAngleCommand(byte angle) {
 		this.angle = angle;
 		this.timestamp = new Date();
 	}
